Make spark app as service. Add a new class for Papp to hold methods related to application and let the main.py free of any logic
